Rename item navigation test files
