+++ Tests/PlayerTests/Playlist/ItemNavigationBackwardChecksTests.swift
@@ -10,17 +10,47 @@ import CoreMedia
 import Nimble
 import PillarboxStreams
 
-final class NavigationSmartBackwardChecksTests: TestCase {
-    func testCanReturnForOnDemandAtBeginningWithoutPreviousItem() {
-        let item = PlayerItem.simple(url: Stream.onDemand.url)
-        let player = Player(item: item)
-        expect(player.streamType).toEventually(equal(.onDemand))
+final class ItemNavigationBackwardChecksTests: TestCase {
+    private static func configuration() -> PlayerConfiguration {
+        .init(navigationMode: .immediate)
+    }
+
+    func testCanReturnToPreviousItem() {
+        let item1 = PlayerItem.simple(url: Stream.onDemand.url)
+        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(items: [item1, item2])
+        player.advanceToNextItem()
         expect(player.canReturnToPreviousItem()).to(beTrue())
+    }
+
+    func testCannotReturnToPreviousItemAtFront() {
+        let item1 = PlayerItem.simple(url: Stream.onDemand.url)
+        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(items: [item1, item2])
+        expect(player.canReturnToPreviousItem()).to(beFalse())
+    }
+
+    func testCannotReturnToPreviousItemWhenEmpty() {
+        let player = Player()
+        expect(player.canReturnToPreviousItem()).to(beFalse())
+    }
+
+    func testWrapAtFrontWithRepeatAll() {
+        let player = Player(item: .simple(url: Stream.onDemand.url))
+        player.repeatMode = .all
+        expect(player.canReturnToPreviousItem()).to(beTrue())
+    }
+
+    func testCannotReturnForOnDemandAtBeginningWithoutPreviousItem() {
+        let item = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(item: item, configuration: Self.configuration())
+        expect(player.streamType).toEventually(equal(.onDemand))
+        expect(player.canReturnToPreviousItem()).to(beFalse())
     }
 
     func testCanReturnForOnDemandNearBeginningWithoutPreviousItem() {
         let item = PlayerItem.simple(url: Stream.onDemand.url)
-        let player = Player(item: item)
+        let player = Player(item: item, configuration: Self.configuration())
         expect(player.streamType).toEventually(equal(.onDemand))
 
         waitUntil { done in
@@ -29,13 +59,13 @@ final class NavigationSmartBackwardChecksTests: TestCase {
             }
         }
 
-        expect(player.canReturnToPreviousItem()).to(beTrue())
+        expect(player.canReturnToPreviousItem()).to(beFalse())
     }
 
     func testCanReturnForOnDemandAtBeginningWithPreviousItem() {
         let item1 = PlayerItem.simple(url: Stream.shortOnDemand.url)
         let item2 = PlayerItem.simple(url: Stream.onDemand.url)
-        let player = Player(items: [item1, item2])
+        let player = Player(items: [item1, item2], configuration: Self.configuration())
         player.advanceToNextItem()
         expect(player.streamType).toEventually(equal(.onDemand))
         expect(player.canReturnToPreviousItem()).to(beTrue())
@@ -44,7 +74,7 @@ final class NavigationSmartBackwardChecksTests: TestCase {
     func testCanReturnForOnDemandNotAtBeginning() {
         let item1 = PlayerItem.simple(url: Stream.shortOnDemand.url)
         let item2 = PlayerItem.simple(url: Stream.onDemand.url)
-        let player = Player(items: [item1, item2])
+        let player = Player(items: [item1, item2], configuration: Self.configuration())
         player.advanceToNextItem()
         expect(player.streamType).toEventually(equal(.onDemand))
 
@@ -60,7 +90,7 @@ final class NavigationSmartBackwardChecksTests: TestCase {
     func testCanReturnForLiveWithPreviousItem() {
         let item1 = PlayerItem.simple(url: Stream.onDemand.url)
         let item2 = PlayerItem.simple(url: Stream.live.url)
-        let player = Player(items: [item1, item2])
+        let player = Player(items: [item1, item2], configuration: Self.configuration())
         player.advanceToNextItem()
         expect(player.streamType).toEventually(equal(.live))
         expect(player.canReturnToPreviousItem()).to(beTrue())
@@ -68,7 +98,7 @@ final class NavigationSmartBackwardChecksTests: TestCase {
 
     func testCannotReturnForLiveWithoutPreviousItem() {
         let item = PlayerItem.simple(url: Stream.live.url)
-        let player = Player(item: item)
+        let player = Player(item: item, configuration: Self.configuration())
         expect(player.streamType).toEventually(equal(.live))
         expect(player.canReturnToPreviousItem()).to(beFalse())
     }
@@ -76,7 +106,7 @@ final class NavigationSmartBackwardChecksTests: TestCase {
     func testCanReturnForDvrWithPreviousItem() {
         let item1 = PlayerItem.simple(url: Stream.onDemand.url)
         let item2 = PlayerItem.simple(url: Stream.dvr.url)
-        let player = Player(items: [item1, item2])
+        let player = Player(items: [item1, item2], configuration: Self.configuration())
         player.advanceToNextItem()
         expect(player.streamType).toEventually(equal(.dvr))
         expect(player.canReturnToPreviousItem()).to(beTrue())
@@ -84,7 +114,7 @@ final class NavigationSmartBackwardChecksTests: TestCase {
 
     func testCannotReturnForDvrWithoutPreviousItem() {
         let item = PlayerItem.simple(url: Stream.dvr.url)
-        let player = Player(item: item)
+        let player = Player(item: item, configuration: Self.configuration())
         expect(player.streamType).toEventually(equal(.dvr))
         expect(player.canReturnToPreviousItem()).to(beFalse())
     }
@@ -92,7 +122,7 @@ final class NavigationSmartBackwardChecksTests: TestCase {
     func testCanReturnForUnknownWithPreviousItem() {
         let item1 = PlayerItem.simple(url: Stream.onDemand.url)
         let item2 = PlayerItem.simple(url: Stream.unavailable.url)
-        let player = Player(items: [item1, item2])
+        let player = Player(items: [item1, item2], configuration: Self.configuration())
         player.advanceToNextItem()
         expect(player.streamType).to(equal(.unknown))
         expect(player.canReturnToPreviousItem()).to(beTrue())
@@ -100,7 +130,7 @@ final class NavigationSmartBackwardChecksTests: TestCase {
 
     func testCannotReturnForUnknownWithoutPreviousItem() {
         let item = PlayerItem.simple(url: Stream.unavailable.url)
-        let player = Player(item: item)
+        let player = Player(item: item, configuration: Self.configuration())
         expect(player.streamType).to(equal(.unknown))
         expect(player.canReturnToPreviousItem()).to(beFalse())
     }
