+++ Tests/PlayerTests/Playlist/ItemNavigationForwardTests.swift
@@ -9,7 +9,58 @@
 import Nimble
 import PillarboxStreams
 
-final class NavigationSmartForwardTests: TestCase {
+final class ItemNavigationForwardTests: TestCase {
+    func testAdvanceToNextItem() {
+        let item1 = PlayerItem.simple(url: Stream.onDemand.url)
+        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(items: [item1, item2])
+        player.advanceToNextItem()
+        expect(player.currentItem).to(equal(item2))
+    }
+
+    func testAdvanceToNextItemAtBack() {
+        let item1 = PlayerItem.simple(url: Stream.onDemand.url)
+        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(items: [item1, item2])
+        player.advanceToNextItem()
+        player.advanceToNextItem()
+        expect(player.currentItem).to(equal(item2))
+    }
+
+    func testAdvanceToNextItemOnFailedItem() {
+        let item1 = PlayerItem.simple(url: Stream.onDemand.url)
+        let item2 = PlayerItem.simple(url: Stream.unavailable.url)
+        let item3 = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(items: [item1, item2, item3])
+        player.advanceToNextItem()
+        player.advanceToNextItem()
+        expect(player.currentItem).to(equal(item3))
+    }
+
+    func testPlayerPreloadedItemCount() {
+        let player = Player(items: [
+            PlayerItem.simple(url: Stream.onDemand.url),
+            PlayerItem.simple(url: Stream.squareOnDemand.url),
+            PlayerItem.simple(url: Stream.mediumOnDemand.url),
+            PlayerItem.simple(url: Stream.onDemand.url),
+            PlayerItem.simple(url: Stream.shortOnDemand.url)
+        ])
+        player.advanceToNextItem()
+
+        let items = player.queuePlayer.items()
+        expect(items).to(haveCount(player.configuration.preloadedItems))
+    }
+
+    func testWrapAtBackWithRepeatAll() {
+        let item1 = PlayerItem.simple(url: Stream.shortOnDemand.url)
+        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(items: [item1, item2])
+        player.repeatMode = .all
+        player.advanceToNextItem()
+        player.advanceToNextItem()
+        expect(player.currentItem).to(equal(item1))
+    }
+
     func testAdvanceForOnDemandWithNextItem() {
         let item1 = PlayerItem.simple(url: Stream.onDemand.url)
         let item2 = PlayerItem.simple(url: Stream.live.url)
