+++ Tests/PlayerTests/Playlist/ItemNavigationSmartBackwardChecksTests.swift
@@ -1,0 +1,107 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+@testable import PillarboxPlayer
+
+import CoreMedia
+import Nimble
+import PillarboxStreams
+
+final class ItemNavigationSmartBackwardChecksTests: TestCase {
+    func testCanReturnForOnDemandAtBeginningWithoutPreviousItem() {
+        let item = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(item: item)
+        expect(player.streamType).toEventually(equal(.onDemand))
+        expect(player.canReturnToPreviousItem()).to(beTrue())
+    }
+
+    func testCanReturnForOnDemandNearBeginningWithoutPreviousItem() {
+        let item = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(item: item)
+        expect(player.streamType).toEventually(equal(.onDemand))
+
+        waitUntil { done in
+            player.seek(at(CMTime(value: 1, timescale: 1))) { _ in
+                done()
+            }
+        }
+
+        expect(player.canReturnToPreviousItem()).to(beTrue())
+    }
+
+    func testCanReturnForOnDemandAtBeginningWithPreviousItem() {
+        let item1 = PlayerItem.simple(url: Stream.shortOnDemand.url)
+        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(items: [item1, item2])
+        player.advanceToNextItem()
+        expect(player.streamType).toEventually(equal(.onDemand))
+        expect(player.canReturnToPreviousItem()).to(beTrue())
+    }
+
+    func testCanReturnForOnDemandNotAtBeginning() {
+        let item1 = PlayerItem.simple(url: Stream.shortOnDemand.url)
+        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(items: [item1, item2])
+        player.advanceToNextItem()
+        expect(player.streamType).toEventually(equal(.onDemand))
+
+        waitUntil { done in
+            player.seek(at(CMTime(value: 5, timescale: 1))) { _ in
+                done()
+            }
+        }
+
+        expect(player.canReturnToPreviousItem()).to(beTrue())
+    }
+
+    func testCanReturnForLiveWithPreviousItem() {
+        let item1 = PlayerItem.simple(url: Stream.onDemand.url)
+        let item2 = PlayerItem.simple(url: Stream.live.url)
+        let player = Player(items: [item1, item2])
+        player.advanceToNextItem()
+        expect(player.streamType).toEventually(equal(.live))
+        expect(player.canReturnToPreviousItem()).to(beTrue())
+    }
+
+    func testCannotReturnForLiveWithoutPreviousItem() {
+        let item = PlayerItem.simple(url: Stream.live.url)
+        let player = Player(item: item)
+        expect(player.streamType).toEventually(equal(.live))
+        expect(player.canReturnToPreviousItem()).to(beFalse())
+    }
+
+    func testCanReturnForDvrWithPreviousItem() {
+        let item1 = PlayerItem.simple(url: Stream.onDemand.url)
+        let item2 = PlayerItem.simple(url: Stream.dvr.url)
+        let player = Player(items: [item1, item2])
+        player.advanceToNextItem()
+        expect(player.streamType).toEventually(equal(.dvr))
+        expect(player.canReturnToPreviousItem()).to(beTrue())
+    }
+
+    func testCannotReturnForDvrWithoutPreviousItem() {
+        let item = PlayerItem.simple(url: Stream.dvr.url)
+        let player = Player(item: item)
+        expect(player.streamType).toEventually(equal(.dvr))
+        expect(player.canReturnToPreviousItem()).to(beFalse())
+    }
+
+    func testCanReturnForUnknownWithPreviousItem() {
+        let item1 = PlayerItem.simple(url: Stream.onDemand.url)
+        let item2 = PlayerItem.simple(url: Stream.unavailable.url)
+        let player = Player(items: [item1, item2])
+        player.advanceToNextItem()
+        expect(player.streamType).to(equal(.unknown))
+        expect(player.canReturnToPreviousItem()).to(beTrue())
+    }
+
+    func testCannotReturnForUnknownWithoutPreviousItem() {
+        let item = PlayerItem.simple(url: Stream.unavailable.url)
+        let player = Player(item: item)
+        expect(player.streamType).to(equal(.unknown))
+        expect(player.canReturnToPreviousItem()).to(beFalse())
+    }
+}
