+++ Tests/PlayerTests/Playlist/ItemNavigationSmartForwardChecksTests.swift
@@ -1,0 +1,72 @@
+//
+//  Copyright (c) SRG SSR. All rights reserved.
+//
+//  License information is available from the LICENSE file.
+//
+
+@testable import PillarboxPlayer
+
+import Nimble
+import PillarboxStreams
+
+final class ItemNavigationSmartForwardChecksTests: TestCase {
+    func testCanAdvanceForOnDemandWithNextItem() {
+        let item1 = PlayerItem.simple(url: Stream.onDemand.url)
+        let item2 = PlayerItem.simple(url: Stream.live.url)
+        let player = Player(items: [item1, item2])
+        expect(player.streamType).toEventually(equal(.onDemand))
+        expect(player.canAdvanceToNextItem()).to(beTrue())
+    }
+
+    func testCannotAdvanceForOnDemandWithoutNextItem() {
+        let item = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(item: item)
+        expect(player.streamType).toEventually(equal(.onDemand))
+        expect(player.canAdvanceToNextItem()).to(beFalse())
+    }
+
+    func testCanAdvanceForLiveWithNextItem() {
+        let item1 = PlayerItem.simple(url: Stream.live.url)
+        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(items: [item1, item2])
+        expect(player.streamType).toEventually(equal(.live))
+        expect(player.canAdvanceToNextItem()).to(beTrue())
+    }
+
+    func testCannotAdvanceForLiveWithoutNextItem() {
+        let item = PlayerItem.simple(url: Stream.live.url)
+        let player = Player(item: item)
+        expect(player.streamType).toEventually(equal(.live))
+        expect(player.canAdvanceToNextItem()).to(beFalse())
+    }
+
+    func testCanAdvanceForDvrWithNextItem() {
+        let item1 = PlayerItem.simple(url: Stream.dvr.url)
+        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(items: [item1, item2])
+        expect(player.streamType).toEventually(equal(.dvr))
+        expect(player.canAdvanceToNextItem()).to(beTrue())
+    }
+
+    func testCannotAdvanceForDvrWithoutNextItem() {
+        let item = PlayerItem.simple(url: Stream.dvr.url)
+        let player = Player(item: item)
+        expect(player.streamType).toEventually(equal(.dvr))
+        expect(player.canAdvanceToNextItem()).to(beFalse())
+    }
+
+    func testCanAdvanceForUnknownWithNextItem() {
+        let item1 = PlayerItem.simple(url: Stream.unavailable.url)
+        let item2 = PlayerItem.simple(url: Stream.onDemand.url)
+        let player = Player(items: [item1, item2])
+        expect(player.streamType).to(equal(.unknown))
+        expect(player.canAdvanceToNextItem()).to(beTrue())
+    }
+
+    func testCannotAdvanceForUnknownWithoutNextItem() {
+        let item = PlayerItem.simple(url: Stream.unavailable.url)
+        let player = Player(item: item)
+        expect(player.streamType).to(equal(.unknown))
+        expect(player.canAdvanceToNextItem()).to(beFalse())
+    }
+}
